--- pscan_unfolded_rev_3.py
@@ -1,193 +1,0 @@
-import math
-
-import torch
-import torch.nn.functional as F
-
-"""
-
-An implementation of the parallel scan operation in PyTorch (Blelloch version).
-This code follows the skeleton proposed by Francois Fleuret in his pscan. However, the keys differences are :
--it has been written in an iterative way (rather than recursive)
--the backward pass has been rewritten
-
-Please see docs/pscan.ipynb for a detailed explanation of what happens here.
-
-"""
-
-def pad_npo2_custom(X):
-    # X : (B, L, D, N)
-
-    # Y : (B, npo2(L), D, N)
-
-    B, L, D, N = X.size()
-
-    len_npo2 = 2**math.ceil(math.log2(L))
-    zeros = torch.zeros(B, len_npo2, D, N, device="cuda")
-    zeros[:L] = X
-    return zeros
-
-class PScan(torch.autograd.Function):
-    @staticmethod
-    def pscan(A, X):
-        # A : (B, D, L, N)
-        # X : (B, D, L, N)
-
-        # modifies X in place by doing a parallel scan.
-        # more formally, X will be populated by these values :
-        # H[t] = A[t] * H[t-1] + X[t] with H[0] = 0
-        # which are computed in parallel (2*log2(T) sequential steps (ideally), instead of T sequential steps)
-        
-        B, D, L, _ = A.size()
-        num_steps = int(math.log2(L))
-
-        # up sweep (last 2 steps unfolded)
-        Aa = A
-        Xa = X
-        for _ in range(num_steps-2):
-            T = Xa.size(2)
-            Aa = Aa.view(B, D, T//2, 2, -1)
-            Xa = Xa.view(B, D, T//2, 2, -1)
-            
-            Xa[:, :, :, 1].add_(Aa[:, :, :, 1].mul(Xa[:, :, :, 0]))
-            Aa[:, :, :, 1].mul_(Aa[:, :, :, 0])
-
-            Aa = Aa[:, :, :, 1]
-            Xa = Xa[:, :, :, 1]
-
-        # we have only 4, 2 or 1 nodes left
-        if Xa.size(2) == 4:
-            Xa[:, :, 1].add_(Aa[:, :, 1].mul(Xa[:, :, 0]))
-            Aa[:, :, 1].mul_(Aa[:, :, 0])
-
-            Xa[:, :, 3].add_(Aa[:, :, 3].mul(Xa[:, :, 2] + Aa[:, :, 2].mul(Xa[:, :, 1])))
-        elif Xa.size(2) == 2:
-            Xa[:, :, 1].add_(Aa[:, :, 1].mul(Xa[:, :, 0]))
-            return
-        else:
-            return
-
-        # down sweep (first 2 steps unfolded)
-        Aa = A[:, :, 2**(num_steps-2)-1:L:2**(num_steps-2)]
-        Xa = X[:, :, 2**(num_steps-2)-1:L:2**(num_steps-2)]
-        Xa[:, :, 2].add_(Aa[:, :, 2].mul(Xa[:, :, 1]))
-        Aa[:, :, 2].mul_(Aa[:, :, 1])
-
-        for k in range(num_steps-3, -1, -1):
-            Aa = A[:, :, 2**k-1:L:2**k]
-            Xa = X[:, :, 2**k-1:L:2**k]
-
-            T = Xa.size(2)
-            Aa = Aa.view(B, D, T//2, 2, -1)
-            Xa = Xa.view(B, D, T//2, 2, -1)
-
-            Xa[:, :, 1:, 0].add_(Aa[:, :, 1:, 0].mul(Xa[:, :, :-1, 1]))
-            Aa[:, :, 1:, 0].mul_(Aa[:, :, :-1, 1])
-
-    @staticmethod
-    def pscan_rev(A, X):
-        # A : (B, D, L, N)
-        # X : (B, D, L, N)
-
-        B, D, L, _ = A.size()
-        num_steps = int(math.log2(L))
-
-        # up sweep (last 2 steps unfolded)
-        Aa = A
-        Xa = X
-        for _ in range(num_steps-2):
-            T = Xa.size(2)
-            Aa = Aa.view(B, D, T//2, 2, -1)
-            Xa = Xa.view(B, D, T//2, 2, -1)
-                    
-            Xa[:, :, :, 0].add_(Aa[:, :, :, 0].mul(Xa[:, :, :, 1]))
-            Aa[:, :, :, 0].mul_(Aa[:, :, :, 1])
-
-            Aa = Aa[:, :, :, 0]
-            Xa = Xa[:, :, :, 0]
-
-        # we have only 4, 2 or 1 nodes left
-        if Xa.size(2) == 4:
-            Xa[:, :, 2].add_(Aa[:, :, 2].mul(Xa[:, :, 3]))
-            Aa[:, :, 2].mul_(Aa[:, :, 3])
-
-            Xa[:, :, 0].add_(Aa[:, :, 0].mul(Xa[:, :, 1].add(Aa[:, :, 1].mul(Xa[:, :, 2]))))
-        elif Xa.size(2) == 2:
-            Xa[:, :, 0].add_(Aa[:, :, 0].mul(Xa[:, :, 1]))
-            return
-        else:
-            return
-
-        # down sweep (first 2 steps unfolded)
-        Aa = A[:, :, 0:L:2**(num_steps-2)]
-        Xa = X[:, :, 0:L:2**(num_steps-2)]
-        Xa[:, :, 1].add_(Aa[:, :, 1].mul(Xa[:, :, 2]))
-        Aa[:, :, 1].mul_(Aa[:, :, 2])
-
-        for k in range(num_steps-3, -1, -1):
-            Aa = A[:, :, 0:L:2**k]
-            Xa = X[:, :, 0:L:2**k]
-
-            T = Xa.size(2)
-            Aa = Aa.view(B, D, T//2, 2, -1)
-            Xa = Xa.view(B, D, T//2, 2, -1)
-
-            Xa[:, :, :-1, 1].add_(Aa[:, :, :-1, 1].mul(Xa[:, :, 1:, 0]))
-            Aa[:, :, :-1, 1].mul_(Aa[:, :, 1:, 0])
-
-    @staticmethod
-    def forward(ctx, A_in, X_in):
-        """
-        Applies the parallel scan operation, as defined above. Returns a new tensor.
-
-        Args:
-            A_in : (B, L, D, N)
-            X_in : (B, L, D, N)
-
-        Returns:
-            H : (B, L, D, N)
-        """
-
-        # pad tensors (and clone btw, which is necessary because of the in-place ops)
-        A = pad_npo2_custom(A_in) # (B, npo2(L), D, N)
-        X = pad_npo2_custom(X_in) # (B, npo2(L), D, N)
-        
-        # prepare tensors
-        A = A.transpose(2, 1) # (B, D, npo2(L), N)
-        X = X.transpose(2, 1) # (B, D, npo2(L), N)
-
-        # parallel scan
-        PScan.pscan(A, X)
-
-        ctx.save_for_backward(A_in, X)
-
-        return X.transpose(2, 1)
-    
-    @staticmethod
-    def backward(ctx, grad_output_in):
-        """
-        Flows the gradient from the output to the input. Returns two new tensors.
-
-        Args:
-            ctx : A_in : (B, L, D, N), X : (B, D, L, N)
-            grad_output_in : (B, L, D, N)
-
-        Returns:
-            gradA : (B, L, D, N), gradX : (B, L, D, N)
-        """
-
-        A_in, X = ctx.saved_tensors
-
-        # prepare tensors
-        A_in = A_in.transpose(2, 1) # (B, D, L, N)
-        A = torch.nn.functional.pad(A_in[:, :, 1:], (0, 0, 0, 1)) # the padding also clones
-
-        grad_output = grad_output_in.transpose(2, 1).clone()
-
-        PScan.pscan_rev(A, grad_output)
-
-        Q = torch.zeros_like(X)
-        Q[:, :, 1:].add_(X[:, :, :-1] * grad_output[:, :, 1:])
-
-        return Q.transpose(2, 1), grad_output.transpose(2, 1)
-    
-pscan = PScan.apply
